requirements.txt
requests == 2.32.3
paho-mqtt == 2.1.0
flask == 2.3.2
